[chore] #34 dropShadow modifier -> common 모듈로 이동
